Updated openstack/openstack Project: openstack/nova  7d3da4d645b878e4eb3c240d9b48124f539eaea7
style-only: s/context/ctx/

Changes the self.context attribute to self.ctx in the
test_resource_provider functional test. Later patches clean up and
condense this test file and having a shorter context attribute allows
us to reduce a number of lines of code in the test file and get rid of
some of the pesky parens on a line by themselves that so many people
hate.

Change-Id: I29c37e260465a574d3134a005dbc14e2a3c6ea00