Updated openstack/openstack Project: openstack/diskimage-builder  b1961e14ea6e3bcdc80ca6e02e80646280b3a86a
Use SELinuxfs to check selinux status

Currently, the cleanup script is using existence of
semanage binary to check if selinux is enabled. However
this is misleading and can lead to problems when selinux
is disabled in a system where the binary exist.

This patch changes the detection logic to use /sys/fs/selinux
directory which is a in-memory filesystem created only when
selinux is really enabled.

Change-Id: I008f8bbc9c8414ce948c601e3907e27764e15a52
Related-Bug: 1706386