Updated openstack/openstack Project: openstack/python-novaclient  2e307bb3e8c92c68e4de7829df60055e9b573c40
Microversion 2.53 - services and hypervisors using UUIDs

Adds support for the 2.53 microversion which changes the
os-services and os-hypervisors APIs to use a UUID for the ID
value on the resource.

Also, the PUT and GET API methods have changed a bit for
both resources in this microversion, so the pythong API
bindings and command lines have been adjusted accordingly.

Part of blueprint service-hyper-uuid-in-api

Change-Id: Ic721143cc154d91e74a8a9dd2c1e991045c94305