Updated openstack/openstack Project: openstack/cinder  6d67e0f15d0e63a4f59526a941f3b93bea111c1f
Add support for OS-Brick force disconnect

OS-Brick disconnect has a force parameter since version 1.15.1 that can
be used to make sure a disconnect call gives more priority to cleaning
up everything than preserving data.

This patch adds support for this new force parameter using it in calls
where we know it's OK to lose data when we encounter an error as long as
we report that we have failed.

Change-Id: Id6a7f4cf264c540473269104daf20f39891b5a1b