Updated openstack/openstack Project: openstack/nova  9d1db10f842286fb59b9316987e075fd8758107d
placement: account for move operations in claim

When a move operation occurs that involves the scheduler, we want the
scheduler to claim resources against the destination host but we do not
want the allocations against the source host to be removed until the
move operation completes. If that were to happen, we would improperly
be freeing resources on the source host before those resources were
actually able to be claimed by another consumer.

So, in this patch, we modify the scheduler report client's
claim_resources() method to first check to see if there are any
allocations for the requested consumer (instance) before attempting to
claim resources against a new set of resource providers. If there *are*
existing allocations, we assume that the scheduler is calling
claim_resources() during a move operation and we adjust the requested
allocation to include the original source host's resources along with
the new destination host's resources.

We ensure that if the original allocation on the source host included
shared resource providers that we do not "double up" the request for
that resource in the new allocation.

Change-Id: Ifd09bd0128ed74347c984a3dcca1304336853865
blueprint: placement-claims
Depends-On: Ib14016a3bc6f2714758ad0291396233218c593c6