Updated openstack/openstack Project: openstack/python-cinderclient  63ac82a55489d55246da939b5ae60b8a65fb9ec7
cinder show with attachments is a mess

The show command currently just dumps the entire attachments
dictionary into the output which is a real mess and completely
screws up the displayed output (shell command).

There's really no reason to do this, we can just give the
attachment ID's and then when you're on the newer versions
you can do an attachment-show for all the crazy details if
you want.  Keep in mind that the list command already shows
the server-id we're attached too, but that might also be
nice from the show command rather than jumping through
multiple commands.  To try and accomodate various use cases
we'll also add an "attached_servers" field to the show command,
but you'll have to coorelate manually from there.

Change-Id: I45ac49d8d9a185c52727c5bc24a6a1323be83689
Closes-Bug: #1494941