Updated openstack/openstack Project: openstack/cinder  93cf04a333c32cf38bcad588533c95ba8580d2aa
Fix no feature to enable 'exclusive-lock' of image in ceph

In normally, it is as default to enable 'exclusive-lock' of image.
But when set 'rbd default features = 1' in 'ceph.conf' file to run
the ceph cluster. The feature is not as default. And there is no
feature to check and enable the 'exclusive-lock' in cinder, which
the 'journaling' depends on it.

Change-Id: Ic11404b7117d38b3e6a32666964bb5a21390821c
Closes-Bug: #1702650