Updated openstack/openstack Project: openstack-infra/tripleo-ci  a0c42cb3eb00c62458bd8f72879f73de4b9839e9
Create a loop device in multinode subnodes

This would be used by ceph-ansible as backend for ceph-osd. For
virtual environments this is solved in OOOQ with the use of "extradisks"
while production environments likely have more than the bootstrap disk
if the node is meant to be used for Ceph.

Note that puppet-ceph supported the local filesystem as backend while
ceph-ansible does not so this wasn't necessary before.

Change-Id: I6ecea0066ef57795030df19a11030ed5be865839