Updated openstack/openstack Project: openstack-infra/project-config  b3e7da6ceda2da966551fa8c8908652ee112032f
Revert "Fix up regex for skipping jobs for certain patrole changes"

This reverts commit fa99152596cf8de43a37826512d269e9b4b089dc.

The reverted commit caused non-patrole jobs with 'gate-tempest-*' to
have their skips to be ignored and thus causing 'gate-tempest-*' jobs
to run when they shouldn't be. For example a doc only change.

A discussion on IRC about this issue can be seen at:
http://eavesdrop.openstack.org/irclogs/%23openstack-infra/%23openstack-infra.2017-07-26.log.html#t2017-07-26T22:42:39

Change-Id: Ie2be8798a6e67f4491bab4376afbb5469ca65caa