Updated openstack/openstack Project: openstack/nova  d38c6f489bda5e6523e13d8a736d70e4ed15ae72
deprecate ``wsgi_log_format`` config variable

Running under a real wsgi environment, such as uwsgi or apache /
mod_wsgi is preferred now, and there is a different logging solution
for that. This was always an eventlet leaky abstraction. So we'll just
deprecate this for now, and the code can be removed once eventlet on
the API side is removed.

Change-Id: I934a1693cbfd539d8575d022519510d19d13aaab