Updated openstack/openstack Project: openstack/python-cinderclient  f5bedfc5bab5bc73f791c0cf2c2c033472b700e1
import content from cli-reference in openstack-manuals

This patch is part of the great doc migration from
openstack-manuals to the individual repos.  The
changes are based on the spec here:

https://specs.openstack.org/openstack/docs-specs/specs/pike/os-manuals-migration.html

The cinder.rst file needs to be renamed index.rst so
that the template that is being used by OpenStack
manuals team will magically find the documentation.

Change-Id: Ia3b5ee2b6aaf17781d3de7546a9477f1c928092f