Updated openstack/openstack Project: openstack-infra/project-config  1c02ab2261afce43bdad6f12625f5a0b377e776a
Change job type for 3nodes job to move to oooq runner

For the new style of tripleo-ci job, we key off the job type in
order to use the correct configuration for the job. This change
will move the 3nodes job to the new style.

This also changes the 3nodes job to run on tripleo-ci and
tripleo-quickstart(-extras) changes. This is essential to make
sure we are checking changes that could break this job.

Change-Id: I429a719b49dce85aa95b4b34040f7646786a54fa