Updated openstack/openstack Project: openstack/horizon  67eba6443900654b5c70501553e65ddb48f33bad
If the flavor_id not in full_flavors, don't make a new API request

Flavors are cached from api.nova.flavor_list(), therefore if we have
machine for which flavor is not in the obtained list, it means the
machine has been created using flavor which is not available anymore.
Additional API calls are just decreasing the performance.

Change-Id: I190ae0bd70fc19a9d2d402a808f465c23c842fa7
Closes-bug: #1655300