Updated openstack/openstack Project: openstack/horizon  cc1196588dedef8853b747eef5e19cdcba51b2ca
Add the Snapshots tab on the Volume Details page

Currently, there is no way to filter and see the snapshots for
just a single Volume.  As the combined list of volume snapshots
increases, so does the difficulty in viewing and using them.  It
would be beneficial to have these snapshots isolated from each other
based on the volume they were created from.

This change adds a "Snapshots" tab to the Volume Details page that
shows a table of snapshots associated with only that specific Volume.
This will not only make it easier to just view the snapshots for a
particular Volume, but will also make it much more approachable to
manage and delete them as well.

Change-Id: I0a80ef2c8171c81f73e3abf5ab461dae3a3a9afe
Closes-Bug: #1323644