Updated openstack/openstack Project: openstack/nova  184411f537493b57a47f824eb7fe5432b683b220
Ensure we unshelve in the cell the instance is mapped

When we create a server it gets put in a specific cell and its
instance_mappings record in the API DB is set to map that instance
to that cell.

When unshelving an instance we want to ensure that the scheduler
only considers compute nodes in the cell where the instance is
mapped.

Change-Id: I96285a1902436cf59b6c6ddf22b69a33503d0f4a
Closes-Bug: #1702935