Updated openstack/openstack Project: openstack/horizon  a956fe7a8bfbdd48cbf9ace8833f171df12baf9e
Allow dropdowns to be in front of div.modal-footer

This removes some oddities with the CSS causing select menus to be
hidden in some circumstances. The side effect of this is that when the
modal body is tall, the window now scrolls instead of the body. I'm not
sure if this is actually bad, but its different.

Change-Id: If890ada4c1c450962789db23e9b4cfc50fe77900
Closes-Bug: 1603496