Updated openstack/openstack Project: openstack-infra/system-config  b9b1c3e8d64d8c650263d765adfda9d1e0dddddd
Fix a few docs table of contents references

While adding the last patch, I noticed the comment from doug about the
entries missing from the TOC that were added to a hidden section. I can
see no reason for these to be hidden, so add them to the main TOC.

Also, there's a warning about an invalid ref in the firehose document.
Fix it.

Change-Id: I86663407356aca0cadd633122a0257ad63d0297c