Updated openstack/openstack Project: openstack/python-novaclient  8dd49ea229ecaf81957372f6cb68ceee5e63a0d3
Add 'Forced down' column in serivce-list

Starting from microversion 2.11,
the 'force_down' parameter is returned in the response of
the "List Compute Services" API (GET /os-services).

But the service-list command does not show the 'forced_down' parameter
in the result table.
So add the 'Forced down' column in the table.

Change-Id: I79e3ba37bdf054604eab40452628dbf30014f6f1
Closes-Bug: #1706486