Updated openstack/openstack Project: openstack/python-novaclient  0a53d19a7feadc329534cf04d74411fc51c29e86
Help text for "--matching" is not clear.

The help text for "hypervisor-list --matching" reads,
"List hypervisors matching the given <hostname>."

This is a bit confusing, since it does talk about
hypervisors as plural, but it still talks about a
singular hostname as opposed to a pattern being matched.

Amended help text to make it clear.

Change-Id: I56fd449001289732254ea99d9b75030834491703
Closes-Bug: #1706525