Updated openstack/openstack Project: openstack/python-novaclient  b47eab6812ab40efd13270b16f707ccf9462dfdb
Allow tuple as for nics value

`novaclient.servers.boot` accepts 'nics' argument with information about
network interfaces to attach. Originally, it accepts a list or a string.

In case if a list value, the proper method iterates over it for
processing and constructs a right structure for Nova-API. A tuple type
is similar to a list type in the processing method and can look more
logical for some developers (a tuple is a sequence of immutable python
object, which can be ideal for predefined nics) and doensn't require
anything to change in code except one validation check.

Also, this patch adds more info in the error message for validation
step.

Change-Id: Ief9d22fefb0a64afc346c56e832c6c0b2b3d8fcb