Updated openstack/openstack Project: openstack-infra/devstack-gate  27e1aebbf90616052397b7d7166a0e1fa7b10bc3
Sync and partprobe after mkpart on ephemeral disk

We are seeing failures (though infrequent) when running mkswap after
mkpart because xvde1 does not exist. Theory is that the partition table
changes have not been recognized by the operating system by the time
mkswap runs so we do an explicit partprobe after syncing changes to
disk.

Related-bug: 1706103
Change-Id: If6aed0c01443324974b293c114239d21fbf5f61e