Updated openstack/openstack Project: openstack-infra/system-config  b18a4a1a398375f53c193c10ab2a828e44183f78
Remove ssh key from github connection and add to executor

The executor needs the github connection config so that it knows how to
interact with the remote system from the merger.

However, if we don't provide an ssh_key to github, it will just use
https for cloning. Since that means less keys to manage and we certainly
aren't cloning from anything private on github - just do that.

Change-Id: I3acb6123398b2c6c3fac45f20a4cfde251862b58