Updated openstack/openstack Project: openstack-infra/zuul  c1c11241d825977abc2cee01bbcca407cbb39454
Fix passing labels to Gerrit when they are not defined in All-Projects

This has been done in Ib591997141b68708bee0b291bb88a93cd3294301, but for
some reason it got rolled back in commit 352264b3.

The --verified or --code-review shortcuts only work when these labels
are defined globally for all projects [1]. The other syntax with a
--label works fine even when e.g. the Verified label is being used just
for a subset of projects.

[1] https://gerrit-review.googlesource.com/#/c/51800/1/Documentation/cmd-review.txt

Change-Id: I2f3a338f18583ac7256b8a2a484cf52be4592fc3