Updated openstack/openstack Project: openstack-infra/system-config  3fb95e67a608a5a723eac9c63175701f7bf2f404
allow redirects in .htaccess files on the static web servers

As part of allowing projects to establish redirects when they move
documentation pages around, we need to let them use .htaccess
files. We only want to allow Redirect and RedirectMatch directives,
though. This change restricts the directives on docs.o.o and adds them
on the static servers used for drafts published during a build (to
allow testing of the redirects).

See
http://lists.openstack.org/pipermail/openstack-dev/2017-July/120240.html
for more details.

Change-Id: I236b7b0a9aae065167bd0aef316603d258e4c3c6
Signed-off-by: Doug Hellmann <doug@doughellmann.com>