Updated openstack/openstack Project: openstack-infra/project-config  68b8f3b77840c8397d2ea482a2635c7765653f8c
Fix swift enable/disable logic in rally-dsmv-* jobs

Desire state is that all rally-dsvm jobs, except py35,
are deploying Swfit. Perviously we have wrong logic for disabling
Swift (it was disable only when it was enabled).

Now we are fixing the behaviour and configurution of all jobs, so jobs
would be configured properly.

Change-Id: I787bdc5f83efc96237dfcdaa0fa7a909baba8956