Updated openstack/openstack Project: openstack-infra/project-config  a037b265e37ed0dcebf840af563d5937596c601e
Update Horizon & D_O_A jobs

This patch updates the jobs to run the appropriate Django versions.
There is some confusion here because of py27 installing Django 1.8 for
Newton and Ocata, and Django 1.11 for Pike. The aim is to have
Django 1.8, 1.9 and 1.10 for Newton / Ocata and Django 1.8, 1.9, 1.10
and 1.11 for Pike.

Change-Id: I90452cff669add8191ca1dd6f6688e26540e9aa4