Updated openstack/openstack Project: openstack/cinder  074babe0c2e9d9bf9942faecf954aaf61a80ba09
Revert "Handle concurrent volume creation request"

This change converted a lock which locked on a
per-image basis to lock on a per-image+volume basis.

This defeats the point of having this lock, which was
to not perform the same image download operation more than
once, consuming unneeded network and storage resources.

This is not an especially strenuous or odd locking
pattern.  If this causes tooz+etcd to fail, there is
likely a bug in tooz or etcd.  There are many places in
Cinder that use locks like this one.

Related-Bug: #1705442

This reverts commit 35ae6fb90380a8e9db936827a4414cb249ef399b.

Change-Id: Ib00fcc61658f6b1e59df66d387f3101ea198d2e7