Updated openstack/openstack Project: openstack/os-net-config  d7a8b39585778e17d995df09079e7a1a0b8780a6
Fix --noop being ignored for dhclient execution

The stop_dhclient_execution() method calls netconfig.execute
directly from the object, instead of from the instance of
the object. This has the effect that when execute is called,
noop is always set false, even if os-net-config was called
with --noop. This results in a failure or network outage.

This change checks the value of self.noop from within the
provider.apply() method before calling the
stop_dhclient_execution method.

Change-Id: Ie2cc4519d7028dceb3a14bd7711839834db0ebdf
Closes-bug: 1699590