Updated openstack/openstack Project: openstack/glance  d24806f0800b8697c90ce85a455374374a62df3d
Create image fails if 'enable_image_import' is set

If user sets 'enable_image_import' to True in glance configuration
file then image creation fails with;
TypeError: 'ImageProxy' object is not callable

The reason is if 'enable_image_import' is True then additional header
is added in the response which was not in proper format. Expected
format is ("header", "value") whereas it was given as
"header: value". Corrected the same to resolve the issue.

Change-Id: Ib762c35d5b53896385cdf5f385ce724c93fc9074