Updated openstack/openstack Project: openstack/ironic  12e495cfa50c93578b3e83f269a47d5067d61382
Removes agent mixin from oneview drivers

This patch removes the agent mixin interface from oneview
drivers and adds a flag based solution instead. Once a set
boot device operation needs to be performed, the operation
is stored to driver_internal_info until the server power
on. Before power the server on again, the operation stored
is performed and then the server is powered on.

Change-Id: I8dabf7ef1ff6e44c062310c5cb3509e545585c0a
Closes-Bug: #1503855