Updated openstack/openstack Project: openstack/oslo.concurrency  683d2682621c23bd2fe5c649bf505fe2f9ccf216
Add debug log to indicate when external lock is taken

In a case when a number of threads are concurrently accessing a
resource that has a external lock there is no indication when
one of the threads has taken the external lock.

This patch adds in a debug log to try and add some clarity.

Change-Id: Ibd7cfd3b39370b1a54f2ab50c75c6b92a876ae1a