Updated openstack/openstack Project: openstack/cinder  12deae0cd2f9890bcbf86e8b07139dfd04d36d01
Add choices to backup_swift_auth

Looking at the sample configuration there is no way to tell what are the
valid options for the "backup_swift_auth" configuration option.

This patch adds to the parameter's definition the list of available
options so it's easier for admins to tell what values are supported.

TrivialFix

Change-Id: I177e8245f9a2af4c9b10f8ce57488a3c12036308