Updated openstack/openstack Project: openstack/ironic  1d1f2d92a8e12ed348443fe99b884af2e1dc9655
Changes log level of a message

This fix changes the log level of a message to 'info'. The message
is related to deletion of swift temporary object when object is not
found on the swift. The swift temporary object being attempted for
delete has a expiration time hence this message need not be of
'warning' level.

Change-Id: I7ac52bd0e1ef61d35722dce352e6bc1844d3d2d2
Closes-Bug: #1629286