Updated openstack/openstack Project: openstack-infra/infra-manual  a8e8b796894ca17b96f1140a312701e36430e11e
Update feature branch creation instructions

The instructions for creating a feature branch have changed
to use the more self-service method through the releases
project yaml file. This updates the 'other-branches' details
for feature branches to give basic info and a pointer to the
more complete README instructions in the releases repo.

Change-Id: I951ebf8172fbfcbbaa21ca469ef6cf022079a736