Updated openstack/openstack Project: openstack/keystone  6d3f29f016f21b760ee778b7519de4497a4fdc56
Add the step to install apache2 libapache2-mod-wsgi

The guide uses the Apache HTTP server with mod_wsgi to serve
identity service, but it did not introduce the step to install
Apache2 and libapache2-mod-wsgi package in this guide. The identity
service will not be started without these two packages.
This patch modify the step 1 in part "Install and configure components"
for installing two packages metioned above.

Change-Id: I8ca55e605f806bdc48f753ab893040d9a76aa93d
Closes-Bug: 1707176