Updated openstack/openstack Project: openstack/glance-specs  fc96fd65a25827cbf2cae53da92ab1939cabb7a2
Eliminate Redundant Downloads of Uncached Images

Multiple requests for an image that is not yet cached on the Glance
API node handling the request currently results in multiple download
requests for the same image from the object store. For example,
1000 concurrent build requests based off an uncached image can result
in 1000 download requests from the object store.

Proposing for `untargeted` as the original author is no longer available
and the spec has been reviewed carefully.

Change-Id: I09c73b4520d5e88a27d9089c709c2e116a2d6ac7
Implements: bp duplicate-downloads