Updated openstack/openstack Project: openstack/heat  2caa46f033469300fb8ca2c79dc5fc61a4229606
Allow CREATE_COMPLETE status when cluster/bay update check

This patch will prevent when check update complete request got earlier
replied than the status change to update in progress in magnum.
This will cause the cluster failed in stack but completed in magnum.
The stack will not be able to take that cluster/bay back at this point.
Partial-Bug: #1702433

Change-Id: I0074b76e6ec925f80bb50a7e67a81cda9438941b