Updated openstack/openstack Project: openstack/nova  07f331be66ffd78edc64462d68cec042d2ff9137
rootwrap.d cleanup mislabeled files

The form for rootwrap.d files is to include a comment before the filter
with the filename (and preferably) the command the filter matches.

This change ensures that these file comments refer to a valid file in the
tree.  They can be checked with something like:
  for i in `awk '/^#.*py *:/ {print $2}' etc/nova/rootwrap.d/compute.filters \
                 | sort -u` ; do
      ls ${i/:}
  done

* I13c701c390784fa1f7809705741abb46e40973be renamed
   .../libvirt/connection.py to .../libvirt/drver.py
* I400db60fcc29c2d5e2d3b9dabc055649138468eb switched to os-brick and
   removed nova/storage/linuxscsi.py
* I5fc2425d2c25076ea87686b2e41be35f66ebb923 moved .../libvirt/volume.py
   into .../libvirt/volume/
* Update one comment to make the awk script above work.
* Add comments as 'chown' and 'tee' are used in nova/virt/libvirt/

Change-Id: I3d89830e4770a7cf88389fac5a2a684554a29bc5