Updated openstack/openstack Project: openstack/nova  9006d8601ecbfe200203cb69e4fba2a346225f10
doc: add FAQ entry for cells v1 config options

It's super confusing for new people to nova and cells
to see config options in a [cells] group and think they
should be changing those, like enable and cell_type.

While we have warnings in the config option help text,
let's also put a reminder in the FAQ page in the docs.

Change-Id: I5e106d9e0743d918e2115d809ac3732c5a3d7a5a