Updated openstack/openstack Project: openstack/tripleo-puppet-elements  ab7ac88006c6c26bbebbe6862792f89d34a65236
Set remove-machine-id to executable

We had previously added an element to remove the machine id from the
images however the script is not executable so it wasn't actually being
run as part of the image building process. This change sets
70-remove-machine-id to be executable so that /etc/machine-id will
actually be removed from the images during the building process.
Additionally this changes the element to clear the file rather than
delete it as if the file is missing, things may not work as expected.

Change-Id: Id3f8c510801f8fcc1fc03415b0ebd23a2ff9da4c
Closes-Bug: #1672361