Updated openstack/openstack Project: openstack/ironic  ccd39bbcb466f33f5a5ee55cc6a0c4461e33974d
Change ramdisk log filename template

ironic collects ramdisk logs from agent, and save it to
/var/log/ironic/deploy.

The time stamp part of log filename is formatted as '%Y-%m-%d-%H:%M:%S',
as an example:

    7b52d13a-9372-448d-9252-fffbaf2db942_1fb52a75-04b6-4c73-80c4-\
    8d49dc30f650_2017-07-06-09:03:18.tar.gz

The colon is not handled properly by tar, it fails to untar a file
with colon in file name.

This patch provides a workaround by changing colon with dash in time
stamp.

Closes-Bug: #1702672
Change-Id: I1d60939768ffa75de050aa309a3203de553ae877