Updated openstack/openstack Project: openstack-infra/infra-manual  49a55ae47fdc3cee8c83e0c7b9f362f83bc6485d
Describe Zuul v3 job naming

This is a proposal on how to name jobs for Zuul v3 to have consistent
names across projects in OpenStack.

It should guide job developers and reviewers.

This is the first step, I expect this document to evolve over time.

This is a followup to Iec35398a1d20a296378002561e38c2934fada63d.

Change-Id: I8ad07230ce3fc39bcaa851c0829c0ee22ff7d908