Updated openstack/openstack Project: openstack/nova  1807d2edb04efba8b1771d5d6eb7ccb1713d8abe
Test resize with placement api

Test cases for resize with confirm and revert scenarios have been
added.

This test depends on the periodic update_available_resources call to heal
the resource allocations. The fake virt driver needed to be tweaked to
work when more than one compute service is created and periodic tasks
are enabled.

The accounting of the current code is very incorrect at the moment, so this
asserts the existing (wrong) behavior, but marks the actual expected
values everywhere so that we can follow up with fixes to remove all of the
annotations.

Change-Id: If7b02bcd8d77e94c7fb42b721792c1391bc0e3b7
blueprint: placement-claims